Enforce template-curly-spacing
Although this change is in contradiction to the commit message from c600bd8b71094ee972b933762d646faef091376d, this change was made to align the template spacing code style to other cases where curly braces occur. For instance, we use spaces for object literals, for import definitions and for jsx props. Why should it be different this time?