feat: Refactor react/jsx-wrap-multilines rules
See https://github.com/peerigon/eslint-config-peerigon/issues/16

BREAKING CHANGE: Always add new line before multiline wraps. This change
was necessary because of the new react/jsx-closing-tag-location
which requires the closing tag to be on the same indentation as the
opening tag. In combination with the parentheses rule, it could lead
to an unstable state where eslint was trying to fix it
by switching back and forth between two states.