feat: Disable arrow-body-style and arrow-parens rule
We should leave the decision up to the developer if parens are necessary or not. This rule can be annoying when it forces the developer to refactor a function when it's obvious that it will contain multiple statements or will have multiple arguments in the future (especially since it can't be fixed automatically).