feat: Disable "class-methods-use-this"
There are cases where this rule interferes with the intended goal.  
Usually classes are used on purpose to share state between functions. However, there are some cases in which functions rightly do not access a state from the class. So in some cases this rule hinders more than it helps.
In short: You can't really forget to access the state when you need it, but if you don't need it, this shouldn't be "criticized".

* Dependency injected parameters (instead of class state)
* Calculating a return result by only calling an imported utility function
* Returning a hardcoded value

e.g.
```ts
	@Get("/status")
	getStatus(): string {
		return "OK";
	}
```